Draft: kaggle download, description, multiprocess
First Draft of the extended version of the tutorial. This commimt is
still WIP and it is going to be released to be tested in Colab (also for
performance)

It includes restructured content and extended description of
dataset and Apache Beam APIs short intro.

Download of the full dataset from Kaggle

API support with Python Multiprocessing to circumvent the GIL
(To be tested for performance on Colab)