Implement the auto-kappa algorithms.
I add some useful functions to backend. Also, because the entry point to algo is in DPEngine but we need to use histograms that are in analysis module, I had to move Histograms and PreAggregateDataExtractors to pipeline_dp module. Extractors are now stored in a separate file due to circular module dependencies too.